Moving deprecation -> warnings
